requirements.txt
numpy>=1.26.0
pandas>=2.2.0
PyYAML>=6.0.1
yfinance>=0.2.54
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.4.0
torch>=2.2.0
tensorboard>=2.16.0
scipy>=1.12.0
